agents-under-hood/Google-Agent-Development-Kit/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-git-assistant"
version = "0.1.0"
authors = [
  { name="Your Name", email="you@example.com" },
]
description = "A CLI tool that uses Google Gemini to automate Git commit messages."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-generativeai",
    "click",
    "python-dotenv",
    "google-adk",
]

[project.scripts]
commit = "ai_git_assistant.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
